Refactor: ReqUser 디렉토리 변경
