+++ src/main/java/com/example/d3nserver/user/controller/UserController.java
@@ -1,6 +1,6 @@
 package com.example.d3nserver.user.controller;
 
-import com.example.d3nserver.auth.jwt.ReqUser;
+import com.example.d3nserver.common.annotation.ReqUser;
 import com.example.d3nserver.common.base.BaseResponse;
 import com.example.d3nserver.quiz.dto.SolvedQuizResponseDto;
 import com.example.d3nserver.quiz.service.SolvedQuizService;
